# Conflicts:
#	CHANGELOG.md
#	Dokan.props
#	appveyor.yml
#	documentations/Doxyfile
#	dokan/access.c
#	dokan/create.c
#	dokan/directory.c
#	dokan/dokan.c
#	dokan/dokan.def
#	dokan/dokan.h
#	dokan/dokan.rc
#	dokan/dokanc.h
#	dokan/lock.c
#	dokan/setfile.c
#	dokan/timeout.c
#	dokan_control/dokanctl.c
#	dokan_fuse/include/dokanfuse.h
#	dokan_fuse/include/fusemain.h
#	dokan_fuse/src/dokanfuse.rc
#	dokan_np/dokan_np.rc
#	dokan_wix/dokanx64.ddf
#	dokan_wix/dokanx86.ddf
#	dokan_wix/version.xml
#	samples/dokan_mirror/mirror.c
#	sys/create.c
#	sys/dokan.h
#	sys/dokan.rc
#	sys/event.c
#	sys/fileinfo.c
#	sys/public.h
#	sys/write.c

+++ dokan_fuse/include/fuse_win.h
@@ -30,8 +30,8 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-int ntstatus_error_to_errno(int win_res);
-int errno_to_ntstatus_error(int err);
+int ntstatus_error_to_errno(long win_res);
+long errno_to_ntstatus_error(int err);
 
 //This stuff is useful only on Windows in MSVC
 #ifdef _MSC_VER
@@ -187,7 +187,7 @@ struct stat64_cygwin
 	blkcnt_t      st_blocks;
 	timestruc_t   st_birthtim;
 };
-/* The following breaks struct stat definiton in native Windows stats.h
+/* The following breaks struct stat definition in native Windows stats.h
 * So whenever referencing st_atime|st_ctime|st_mtime, replacing is needed.
 */
 /*
